Resolve a port to the processes listening on it
Log monitoring can already follow a process's stdout and stderr given file
descriptors, but nothing could find those descriptors for a dev server that is
merely running on a port. This is the first half of that: port to pid. Reading
the descriptors themselves comes later, and only some of the cases are
recoverable at all.

Adds `PortNumber`, a value type constrained to 1...65535. Port 0 is the
"assign me one" sentinel for bind and never an address anything is reachable
on, so admitting it would leave every consumer re-checking for it. The type
also carries the two conversions that are easy to get wrong -- network byte
order, and the hex in /proc/net/tcp -- so they are unit-testable away from the
scans that use them.

`ListeningProcessLocator` resolves a port to the processes holding a listening
socket on it, natively on both platforms: /proc/net/tcp{,6} plus an fd walk on
Linux, libproc on Darwin. It is a sibling of `PortProbe` rather than a
replacement. A probe answers "is anything accepting connections", which stays
true for other users, other namespaces and remote hosts; a locator answers
"which of my processes owns this", and returns nothing in all three of those
cases. `TailscaleBinder` preflight therefore still uses the probe.

The locator throws where the probe returns false, because "nothing is
listening" and "the scan could not run" need different messages once this
reaches the CLI.

Tests run ungated on both platforms. The fixture that matters is
`SocketHoldingProcess`: it binds a listener, hands the descriptor to a child as
fd 0, then closes the parent's copy, so the port is attributable to the child
alone. Resolving your own process passes even for an implementation that only
ever inspects getpid(); this does not.

+++ Tests/TailregCoreTests/Tailscale/IntegrationTests.swift
@@ -21,12 +21,12 @@ struct `Tailscale integration tests` {
     defer { listener.stop() }
 
     let binding = try await binder.bind(
-      localPort: listener.port,
+      localPort: listener.port.intValue,
       to: .explicit(Self.integrationPort)
     )
 
     #expect(binding.tailnetPort == Self.integrationPort)
-    #expect(binding.localPort == listener.port)
+    #expect(binding.localPort == listener.port.intValue)
     #expect(binding.isManaged)
     #expect(binding.hostname.hasSuffix(".ts.net"))
 
@@ -40,7 +40,7 @@ struct `Tailscale integration tests` {
   func `Refuses A Local Port With Nothing Listening`() async throws {
     let temp = try TempDirectory()
     let listener = try LoopbackListener()
-    let deadPort = listener.port
+    let deadPort = listener.port.intValue
     listener.stop()
 
     await #expect(throws: TailscaleError.noLocalServerListening(port: deadPort)) {
@@ -61,12 +61,12 @@ struct `Tailscale integration tests` {
     defer { Task { try? await binder.unbind(tailnetPort: Self.integrationPort) } }
 
     try await binder.bind(
-      localPort: alpha.port,
+      localPort: alpha.port.intValue,
       to: .explicit(Self.integrationPort),
       mountPath: "/alpha"
     )
     try await binder.bind(
-      localPort: beta.port,
+      localPort: beta.port.intValue,
       to: .explicit(Self.integrationPort),
       mountPath: "/beta"
     )
@@ -74,10 +74,10 @@ struct `Tailscale integration tests` {
     let both = try await binder.bindings().filter { $0.tailnetPort == Self.integrationPort }
     #expect(both.map(\.mountPath).sorted() == ["/alpha", "/beta"])
 
-    try await binder.unbind(localPort: alpha.port)
+    try await binder.unbind(localPort: alpha.port.intValue)
 
     let survivors = try await binder.bindings().filter { $0.tailnetPort == Self.integrationPort }
     #expect(survivors.map(\.mountPath) == ["/beta"])
-    #expect(survivors.map(\.localPort) == [beta.port])
+    #expect(survivors.map(\.localPort) == [beta.port.intValue])
   }
 }
